ST-2192: Add whitelist-commit-text-for-ours to the merge command
Problem:

We want to automate the pint merge with CI pipeline. However in-order to
do that, we would want to have a way to handle special commits like ci
skip, or bump version by changing the merge strategy dynamically to
ours

Solution:

We will add a new property called whitelist-commit-text-for-ours to the pint
merge command, which will inspect each unmerged commit, check if the
commit message contains the whitelisted commits and if present, it will
use a ours merge strategy for that commit alone.

Commit changes:
- Modified merge.py file to accept the newly added
  whitelist-commit-text-for-ours option
- Merge will now work at commit level rather than at the head level,
  where we inspect each commit for the commit message and run merge with
the appropriate merge strategy.